refactor: Creates Factory for Immutable class InformationFields.[#151650459]
